+++ src/lav18/unidubna/controller/Controller.java
@@ -1,0 +1,7 @@
+package lav18.unidubna.controller;
+
+import java.util.Scanner;
+
+public class Controller {
+
+}
